refactor(profile): Remove profile/edit page
