--- app/(route)/dashboard/profile/edit/page.tsx
@@ -1,16 +1,0 @@
-import ProfileEdit from '@/pages/dashboard/ProfileEdit'
-import { NextPage } from 'next'
-
-export const metadata = {
-  title: 'Edit Profile'
-}
-
-const EditProfilePage: NextPage = () => {
-  return (
-    <>
-      <ProfileEdit />
-    </>
-  )
-}
-
-export default EditProfilePage
